sw/requirements.txt
-c constraints.txt

pyserial
intelhex
argh
